Factor format string highlighting out
